refactor(http): split resources.rs into thin facade + sibling modules
Break the 729-line `crates/dcc-mcp-http/src/resources.rs` module into
focused siblings using the `#[path]`-included submodule pattern
established by the file-logging / job / watcher / action-chain
refactors:

- `resources_types.rs` — the `ResourceProducer` trait, `ProducerContent`,
  `ResourceError`, `ResourceResult`, and the `uri_scheme` helper.
- `resources_producers.rs` — the four built-in producers (`SceneProducer`,
  `CaptureProducer`, `AuditProducer`, `ArtefactProducer`) plus the
  `parse_audit_limit` query-string helper.
- `resources_tests.rs` — the full MCP Resources unit + `tokio::test`
  suite (11 cases) moved out of the main file.

`resources.rs` is now a thin facade that keeps only the
`ResourceRegistry` public type (producer wiring, scene snapshot,
subscription state, `notify_updated` fan-out, audit-log forwarding)
and `pub use` re-exports. No public API changes.

Documentation: `docs/guide/architecture.md` now records the new
layout under the `dcc-mcp-http` section.
